avoid implicit use of abort() which fails at link time
So, when USB stuff is disabled (I first noticed it when building for
fx-CP and right now when trying to disable USB with a config), abort()
is not linked automatically into the program, and these references from
UI code end up failing to link.

The reference likely appears because jlayout_get_*() return NULL if the
actual layout doesn't match the request, and the compiler wants to
protect against NULL dereferencing (I'm not sure if it does that all the
time).

_abort is of course available in fxlibc, I suspect there is another
layer of ld failing to rescan fat libraries after LTO, a bug that has
been fixed in the latest versions. So I won't pursue this for now,
hoping it will go away with the next binutils update.